requirements.txt
discord.py
squarecloud-api
python-dotenv

